Move test_image_to_tensorflow
